Add files via upload
If you run "./logic.exe" it WILL NOT RUN because the file paths are formatted for the linux server. Everything on here is here to simply show what is found on the linux server. The only file that will run here is "./login.exe" which will help you to login to the server and run the code on there.