Add files via upload
All this code will not work if you run it. It is just there to show you what is on the server and the backend of the website. The ONLY code that wil run here is "./login.exe" which is there to prompt you to do the login as explained in the README.